feat: select multiple panels
