+++ apps/management-app/src/components/Authenticated/Officer/LocationList/index.tsx
@@ -1,15 +1,7 @@
 import {
   Box,
-  Checkbox,
-  FormControl,
-  InputLabel,
-  ListItemText,
-  MenuItem,
-  OutlinedInput,
   Pagination,
   Paper,
-  Select,
-  SelectChangeEvent,
   Table,
   TableBody,
   TableCell,
@@ -22,10 +14,10 @@ import { useAppSelector } from '@/store';
 import CenterLoading from '@/components/Common/CenterLoading';
 import { Edit, Info } from '@/components/Common/Icons';
 import { ListWrapper } from '@/components/Common/Layout/ScreenWrapper';
+import WardSelect from '@/components/Common/WardSelect';
 import { LocationStatus } from '@/constants/location';
 import { useGetLocationsQuery } from '@/store/api/officer/locationApiSlice';
-import { useGetWardsByDistrictIdQuery } from '@/store/api/officer/wardApiSlide';
-import { Location, Ward } from '@/types/officer-management';
+import { Location } from '@/types/officer-management';
 import { formatDateTime } from '@/utils/datetime';
 import { capitalize } from '@/utils/format-string';
 
@@ -41,74 +33,6 @@ const titles = [
   'Status',
   '',
 ];
-
-const ITEM_HEIGHT = 48;
-const ITEM_PADDING_TOP = 8;
-const MenuProps = {
-  PaperProps: {
-    style: {
-      maxHeight: ITEM_HEIGHT * 4.5 + ITEM_PADDING_TOP,
-      width: 200,
-    },
-  },
-};
-
-function WardSelect({
-  wards,
-  setWards,
-}: {
-  wards: number[];
-  setWards: (wards: number[]) => void;
-}) {
-  const districtId = '18';
-
-  const [allWards, setAllWards] = useState<Ward[]>([]);
-
-  const { data } = useGetWardsByDistrictIdQuery(districtId);
-
-  useEffect(() => {
-    if (data) {
-      setAllWards(data);
-    }
-  }, [data]);
-
-  const handleChange = (event: SelectChangeEvent<number[]>) => {
-    const {
-      target: { value },
-    } = event;
-
-    setWards(
-      typeof value === 'string' ? value.split(',').map((id) => +id) : value,
-    );
-  };
-
-  return (
-    <FormControl sx={{ width: 250, marginBottom: 2 }}>
-      <InputLabel id="select-wards">Wards</InputLabel>
-      <Select
-        labelId="select-wards"
-        id="select-wards"
-        multiple
-        value={wards}
-        onChange={handleChange}
-        input={<OutlinedInput label="Wards" />}
-        renderValue={(selectedIds) =>
-          selectedIds
-            .map((id) => allWards.find((ward) => ward.id === id)?.name)
-            .join(', ')
-        }
-        MenuProps={MenuProps}
-      >
-        {allWards.map((ward) => (
-          <MenuItem key={ward?.id} value={ward?.id}>
-            <Checkbox checked={wards.indexOf(ward?.id) > -1} />
-            <ListItemText primary={ward?.name} />
-          </MenuItem>
-        ))}
-      </Select>
-    </FormControl>
-  );
-}
 
 export default function LocationList() {
   const role = useAppSelector((state) => state.user?.profile?.role);
