+++ apps/management-app/src/components/Authenticated/Officer/PanelList/index.tsx
@@ -13,6 +13,7 @@ import { useEffect, useState } from 'react';
 import CenterLoading from '@/components/Common/CenterLoading';
 import { Edit, Info, Response } from '@/components/Common/Icons';
 import { ListWrapper } from '@/components/Common/Layout/ScreenWrapper';
+import WardSelect from '@/components/Common/WardSelect';
 import { PanelStatus } from '@/constants/panel';
 import { useGetPanelsQuery } from '@/store/api/officer/panelApiSlide';
 import { Panel } from '@/types/officer-management';
@@ -35,15 +36,17 @@ const titles = [
 export default function PanelList() {
   const [page, setPage] = useState<number>(1);
   const [panels, setPanels] = useState<Panel[]>([]);
+  const [wards, setWards] = useState<number[]>([]);
 
   const { data, isLoading, refetch } = useGetPanelsQuery({
     page: page,
     take: 10,
+    wards: wards,
   });
 
   useEffect(() => {
     refetch();
-  }, [panels, refetch]);
+  }, [panels, page, refetch, wards]);
 
   useEffect(() => {
     if (data) {
@@ -57,69 +60,82 @@ export default function PanelList() {
 
   return (
     <ListWrapper label="Panels" btnLabel="Create" btnLink={'/panels/create'}>
-      <TableContainer component={Paper}>
-        <Table sx={{ minWidth: 650 }} aria-label="panels">
-          <TableHead>
-            <TableRow>
-              {titles.map((title) => (
-                <TableCell align="center" key={title}>
-                  {title}
-                </TableCell>
-              ))}
-            </TableRow>
-          </TableHead>
-          <TableBody>
-            {panels?.length !== 0 ? (
-              panels.map((panel: Panel) => (
-                <TableRow key={panel?.id}>
-                  <TableCell align="center">{panel?.id}</TableCell>
-                  <TableCell align="center">{panel?.location?.name}</TableCell>
-                  <TableCell align="center">
-                    {panel?.location?.fullAddress}
-                  </TableCell>
-                  <TableCell align="center">
-                    {panel?.location?.ward?.name}
-                  </TableCell>
-                  <TableCell align="center">
-                    {panel?.location?.district?.name}
-                  </TableCell>
-                  <TableCell align="center">{panel?.companyEmail}</TableCell>
-                  <TableCell align="center">
-                    {formatDateTime(panel?.createdAt)}
-                  </TableCell>
-                  <TableCell align="center">
-                    {formatDateTime(panel?.updatedAt)}
-                  </TableCell>
-                  <TableCell align="center">
-                    {capitalize(panel?.status)}
-                  </TableCell>
-                  <TableCell>
-                    <Box
-                      sx={{
-                        display: 'flex',
-                        gap: 2,
-                      }}
-                    >
-                      <Info link={`/panels/${panel?.id}`} />
-                      <Edit link={`/panels/${panel?.id}/edit`} />
+      <Box
+        sx={{
+          width: '100%',
+        }}
+      >
+        <WardSelect wards={wards} setWards={setWards} />
 
-                      {panel?.status === PanelStatus?.DRAFT && (
-                        <Response link={`/panels/${panel?.id}/send-request`} />
-                      )}
-                    </Box>
+        <TableContainer component={Paper}>
+          <Table sx={{ minWidth: 650 }} aria-label="panels">
+            <TableHead>
+              <TableRow>
+                {titles.map((title) => (
+                  <TableCell align="center" key={title}>
+                    {title}
+                  </TableCell>
+                ))}
+              </TableRow>
+            </TableHead>
+            <TableBody>
+              {panels?.length !== 0 ? (
+                panels.map((panel: Panel) => (
+                  <TableRow key={panel?.id}>
+                    <TableCell align="center">{panel?.id}</TableCell>
+                    <TableCell align="center">
+                      {panel?.location?.name}
+                    </TableCell>
+                    <TableCell align="center">
+                      {panel?.location?.fullAddress}
+                    </TableCell>
+                    <TableCell align="center">
+                      {panel?.location?.ward?.name}
+                    </TableCell>
+                    <TableCell align="center">
+                      {panel?.location?.district?.name}
+                    </TableCell>
+                    <TableCell align="center">{panel?.companyEmail}</TableCell>
+                    <TableCell align="center">
+                      {formatDateTime(panel?.createdAt)}
+                    </TableCell>
+                    <TableCell align="center">
+                      {formatDateTime(panel?.updatedAt)}
+                    </TableCell>
+                    <TableCell align="center">
+                      {capitalize(panel?.status)}
+                    </TableCell>
+                    <TableCell>
+                      <Box
+                        sx={{
+                          display: 'flex',
+                          gap: 2,
+                        }}
+                      >
+                        <Info link={`/panels/${panel?.id}`} />
+                        <Edit link={`/panels/${panel?.id}/edit`} />
+
+                        {panel?.status === PanelStatus?.DRAFT && (
+                          <Response
+                            link={`/panels/${panel?.id}/send-request`}
+                            title="Send Request"
+                          />
+                        )}
+                      </Box>
+                    </TableCell>
+                  </TableRow>
+                ))
+              ) : (
+                <TableRow>
+                  <TableCell align="center" colSpan={9}>
+                    No results
                   </TableCell>
                 </TableRow>
-              ))
-            ) : (
-              <TableRow>
-                <TableCell align="center" colSpan={9}>
-                  No results
-                </TableCell>
-              </TableRow>
-            )}
-          </TableBody>
-        </Table>
-      </TableContainer>
+              )}
+            </TableBody>
+          </Table>
+        </TableContainer>
+      </Box>
 
       <Pagination
         count={data?.totalPages}
