+++ apps/management-app/src/store/api/officer/panelApiSlide.ts
@@ -14,21 +14,35 @@ export const officerPanelApiSlice = apiSlice.injectEndpoints({
       {
         page?: number;
         take?: number;
-        wards?: string;
+        wards?: number[];
         districts?: string;
         typeId?: number;
       }
     >({
-      query: (arg) => ({
-        url: `/panels`,
-        params: {
-          page: arg.page,
-          take: arg.take,
-          wards: arg.wards,
-          districts: arg.districts,
-          typeId: arg.typeId,
-        },
-      }),
+      query: (arg) => {
+        if (arg.wards && arg?.wards?.length > 0) {
+          return {
+            url: `/panels`,
+            params: {
+              page: arg.page,
+              take: arg.take,
+              wards: arg.wards.join(','),
+              districts: arg.districts,
+              typeId: arg.typeId,
+            },
+          };
+        }
+
+        return {
+          url: `/panels`,
+          params: {
+            page: arg.page,
+            take: arg.take,
+            districts: arg.districts,
+            typeId: arg.typeId,
+          },
+        };
+      },
     }),
     getPanelById: build.query<Panel, string>({
       query: (id) => `/panels/${id}`,
